evaluations
outputs accuracy score, confusion matrix and classification report. Not
fully complete since there is a case that i don’t know how to handle
yet. Need to know how to handle truth boxes > minHeigh/minWidth but
also have a real box in the photo. In test case this comes up as a true
negative but it might really not be.